wsf: Handle relative imports and import module.wsf for main
Imports prefixed with @ are located in the search path (currently only
wslib). Otherwise, imports are relative to the importing file. If a
module.wsf file exists adjacent to the main file, it is automatically
imported without a scope (i.e., those labels can be used without
importing and prefixing with the module name).